UI - Attendance Sheet
ALPE = Absent, Late, Present, Excused

Not yet added: Trying to figure out how to change the color of
individual cells or rows of a table.
When the user selects a cell, and chooses one of the ALPE buttons, that
cell/row should change color into the color of the ALPE chosen.